Add DtypeOptimizer when solving steps

+++ rubiks_cube/configuration/types.py
@@ -5,8 +5,8 @@ import numpy.typing as npt
 
 # Arrays representing Rubik's cube masks, patterns, and permutations
 CubeMask: TypeAlias = npt.NDArray[np.bool_]
-CubePattern: TypeAlias = npt.NDArray[np.int_]
-CubePermutation: TypeAlias = npt.NDArray[np.int_]
+CubePattern: TypeAlias = npt.NDArray[np.uint]
+CubePermutation: TypeAlias = npt.NDArray[np.uint]
 CubeColor: TypeAlias = npt.NDArray[np.str_]
 CubeState: TypeAlias = CubeMask | CubePattern | CubePermutation | CubeColor
