+++ rubiks_cube/solver/old_bidirectional_solvers.py
@@ -115,7 +115,7 @@ def bidirectional_solver_v1(
     }
 
     # Last searched permutations and all searched states on inverse permutation
-    identity = np.arange(initial_permutation.size)
+    identity = np.arange(initial_permutation.size, dtype=initial_permutation.dtype)
     solved_str = encode(identity, pattern)
     last_states_inverse: dict[str, tuple[CubePattern, list[str]]] = {solved_str: (identity, [])}
     searched_states_inverse: dict[str, tuple[CubePattern, list[str]]] = {solved_str: (identity, [])}
